Consolidated extract files to single file
